strip out stuff not needed for lite version
for some reason, the eventing system causes the system to crash
on the lite version. I think there are some floating inputs that
are causing interrupts to trigger like crazy and I can't figure out
how to stop that. So, I simply turned that off and rewrote the system
to use polling; it's much more reliable.